Se trabajó sobre las tareas asignadas.
Descripción: Se realizaron las validaciones en Docente y DocenteDTO.
Adicionalmente se actualizaron los Services y el Controller de Docente.

Autor: Rios, Mario

+++ tp4_PeñalozaAbel/src/main/java/ar/edu/unju/fi/repository/DocenteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import ar.edu.unju.fi.model.Docente;
 
 @Repository
-public interface DocenteRepository extends JpaRepository<Docente,Integer> {
+public interface DocenteRepository extends JpaRepository<Docente,Long> {
 
 }
